% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-.R
\name{cran_error_handler}
\alias{cran_error_handler}
\alias{http_condition}
\title{Creates and handle a (error) condition}
\usage{
cran_error_handler(req, res, e)

http_condition(status_code, message, ..., type = NULL, call = sys.call(-1))
}
\arguments{
\item{req, res}{A "request"- and "response"-object, respectively}

\item{e}{The error/[condition][base::conditions] that was thrown by a handler,
preferable a [http_condition].}

\item{status_code}{Integer HTTP Status Code}

\item{message}{Optional message to display (text).}

\item{...}{Other things to include in exception.}

\item{type}{Type of condition to generate.
Must be one of `error`, `warning` or `message`.}

\item{call}{The call stored in the condition object.}
}
\value{
`NULL` invisibly; called to modify the response.
}
\description{
Creates and handle a (error) condition
}
\keyword{internal}
