% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-static.R
\name{directory_listing}
\alias{directory_listing}
\alias{directory_listing_html}
\title{Directory listing}
\usage{
directory_listing(req, res, path, path_prefix)

directory_listing_html(path, url_path, add_dir = TRUE)
}
\arguments{
\item{req, res}{A "request"- and "response"-object, respectively}

\item{path}{Path to directory to list}

\item{path_prefix}{Optional URL-component, if the repository exists at a
subdirectory of the host (see \link{run}'s \code{path}). Here, it is only for
decorative purposes.}

\item{url_path}{The entire, relative path, that is displayed to
end user in the URL.}

\item{add_dir}{Logical, if the requested URL does not end with a '/',
set this to TRUE, else all links will point to the parent directory.}
}
\value{
\code{directory_listing} returns the response-object.
\code{directory_listing_html} returns a HTML-string with the entire page.
}
\description{
Creates a simple HTML page with table of all files and subdirectories.
May throw a 403 or 404 code.
}
