% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-static.R
\name{extract_cran_path_parts}
\alias{extract_cran_path_parts}
\title{Splits repository "contrib"-paths into a directory path and filename.}
\usage{
extract_cran_path_parts(s)
}
\arguments{
\item{s}{The requested path, e.g. \code{src/contrib} or
\code{bin/windows/contrib/4.0/my_package_v0.1.0.zip}.}
}
\value{
A character vector of length 2, with directory path and filename.
If \code{s} was a directory (e.g., "src/contrib", "src/contrib/",
"bin/windows/contrib/4.0", etc.) the second element is "".
Can return \code{NULL} if \code{s} does not fit the repository directory structure.
}
\description{
Splits the requested path into directory-part and filename-part,
while ensuring some security to where it may point to.
}
\keyword{internal}
