% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microCRAN-package.R
\docType{package}
\name{microCRAN-package}
\alias{microCRAN}
\alias{microCRAN-package}
\title{microCRAN: Hosting an Independent CRAN Repository}
\description{
Stand-alone HTTP capable R-package repository, that fully supports R's install.packages() and available.packages(). It also contains API endpoints for end-users to add/update packages. This package can supplement 'miniCRAN', which has functions for maintaining a local (partial) copy of 'CRAN'. Current version is bare-minimum without any access-control or much security.
}
\section{CRAN directory structure}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{/root
/---/src/contrib/
/---/src/contrib/PACKAGES(?.rds|.gz)
/---/src/contrib/microCRAN_v1.0.0.tar.gz
/---/src/contrib/4.0/PACKAGES
/---/src/contrib/4.1/PACKAGES
/---/src/contrib/4.2/PACKAGES
/---/src/contrib/4.3/PACKAGES
/---/bin/windows/contrib/4.0/PACKAGES
/---/bin/windows/contrib/4.0/microCRAN_v1.0.0.zip
/---/bin/windows/contrib/4.1/PACKAGES
/---/bin/windows/contrib/4.2/PACKAGES
/---/bin/windows/contrib/4.3/PACKAGES
/---/bin/macosx/contrib/4.0/PACKAGES
/---/bin/macosx/contrib/4.0/microCRAN_v1.0.0.tgz
/---/bin/macosx/contrib/4.1/PACKAGES
/---/bin/macosx/contrib/4.2/PACKAGES
/---/bin/macosx/contrib/4.3/PACKAGES
}\if{html}{\out{</div>}}
}

\author{
\strong{Maintainer}: Stefan McKinnon Edwards \email{sme@iysik.com} (\href{https://orcid.org/0000-0002-4628-8148}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Kamstrup A/S [copyright holder]
}

}
