% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-static.R
\name{path_is_safe_for_cran}
\alias{path_is_safe_for_cran}
\title{Test if path matches repository directory structure}
\usage{
path_is_safe_for_cran(s)
}
\arguments{
\item{s}{The path to check, typically `req$PATH_INFO` in a request handler.}
}
\value{
Logical
}
\description{
Tests if the requested resource maps to a safe path in the respository,
i.e., that it points to `/src/contrib`, `/bin/windows/contrib` or
`/bin/macosx/contrib`, and does not attempt to weasle back out with `..`.
}
\keyword{internal}
