% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers-.R
\name{pr_add_package}
\alias{pr_add_package}
\title{Plumber route for adding package to repository}
\usage{
pr_add_package(pr, path = "/add", repo_dir)
}
\arguments{
\item{pr}{A Plumber router-object}

\item{path}{The path to the endpoint}

\item{repo_dir}{Path to local directory, where the root of the repository is.
The (source) packages will be stored locally at
\verb{\{repo_dir\}/src/contrib/}.}
}
\value{
\code{NULL} invisibly; called to modify the response.
}
\description{
Creates a \link[plumber:pr_handle]{Plumber route} that handles an incoming
R-package.
Use rather \code{\link[=build]{build()}} to build the entire API.
}
\seealso{
\code{\link[=build]{build()}}, \link[plumber:pr_handle]{plumber::pr_post}
}
