% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{r_package_type}
\alias{r_package_type}
\title{Matches package file extension to package type}
\usage{
r_package_type(fn)
}
\arguments{
\item{fn}{Filename or path}
}
\value{
"source", "mac.binary" or "win.binary", depending on file extension,
or throws a \link{http_condition}.
}
\description{
Matches package file extension to package type
}
\keyword{internal}
