% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{read_DESCRIPTION_zip}
\alias{read_DESCRIPTION_zip}
\alias{read_DESCRIPTION_tar}
\title{Read DESCRIPTION file from package}
\usage{
read_DESCRIPTION_zip(fn)

read_DESCRIPTION_tar(fn)
}
\arguments{
\item{fn}{Path to either zip, tar.gz or tgz file.}
}
\value{
List-object with contents of DESCRIPTION file.
}
\description{
Read DESCRIPTION file from package
}
\examples{
package <- system.file('extdata/microCRAN_0.1.0.zip',
  package='microCRAN', mustWork=TRUE)
read_DESCRIPTION_zip(package)

}
